Update tests for client Id get/set
